search_agent_project/requirements.txt
google-adk>=1.0.0
google-genai>=1.16.1
google-cloud-aiplatform>=1.94.0
python-dotenv>=1.0.0
